added TableDirector file
